requirement.txt
absl-py==2.3.1
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.8.6
aiohttp-cors==0.7.0
aiosignal==1.3.2
alabaster==0.7.16
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asn1crypto==1.5.1
asttokens==3.0.0
async-lru==2.0.5
async-timeout==4.0.3
attrs==25.3.0
audioread==3.0.1
Babel==2.15.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blessed==1.21.0
blinker==1.9.0
borb==2.1.25
boto3==1.38.18
botocore==1.38.18
cachetools==5.5.2
certifi==2024.7.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
cli_exit_tools==1.2.7
click==8.1.8
cloudpickle==3.1.1
colorful==0.5.6
comm==0.2.3
contourpy==1.3.0
cramjam==2.10.0
crcmod==1.7
crypto==1.4.1
cryptography==43.0.3
cssselect==1.3.0
cycler==0.12.1
Cython==3.1.0
dashscope==1.23.3
dbus-python==1.2.16
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
devscripts==2.21.3+deb11u1
dill==0.4.0
distlib==0.3.9
distro==1.9.0
distro-info==1.0+deb11u1
dnspython==2.7.0
docker-pycreds==0.4.0
docutils==0.19
easyocr==1.7.2
ecdsa==0.19.1
einops==0.8.0
entrypoints==0.4
enum34==1.1.10
et_xmlfile==2.0.0
eval_type_backport==0.2.2
evdev==1.9.2
exceptiongroup==1.3.0
executing==2.2.0
fabric==3.2.2
Farama-Notifications==0.0.4
fastapi==0.98.0
fastdtw==0.3.4
fastjsonschema==2.21.2
filelock==3.15.4
fire==0.7.0
flash-attn==2.3.3
Flask==3.0.3
fonttools==4.58.0
formulas==1.2.9
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2024.6.1
ftfy==6.3.1
func-timeout==4.3.5
gdown==5.2.0
gevent==24.11.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
gpustat==1.0.0
greenlet==3.2.2
groq==0.25.0
grpcio==1.59.5
gunicorn==20.1.0
gymnasium==0.28.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.31.2
idna==3.7
ImageHash==4.3.2
imageio==2.37.0
imagesize==1.4.1
importlib_metadata==8.2.0
importlib_resources==6.5.2
invoke==2.2.0
iotop==0.6
ipaddress==1.0.23
ipykernel==6.29.5
ipython==8.18.1
ipython-genutils==0.2.0
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jax-jumpy==1.0.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.0
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter==1.0.0
jupyter_client==7.4.9
jupyter-console==6.6.3
jupyter_core==5.8.1
jupyter-events==0.12.0
jupyter-kernel-gateway==3.0.1
jupyter-lsp==2.2.6
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.7
lark==1.2.2
lazy_loader==0.4
lib-detect-testenv==2.0.8
librosa==0.11.0
linkify-it-py==2.0.3
llvmlite==0.43.0
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.17.2
merlin_kernel==0.1
mistune==3.1.3
mlx-python-sdk==0.3.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.4.3
multiprocess==0.70.18
mutagen==1.47.0
Naked==0.1.32
nbclassic==1.3.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.4
none==0.1.1
notebook==6.5.7
notebook_shim==0.2.4
numba==0.60.0
numpy==1.24.4
numpy-financial==1.0.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==11.495.46
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.1.105
oauth2client==4.1.3
odfpy==1.4.1
openai==1.51.2
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python==4.8.1.78
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
overrides==7.7.0
packaging==24.1
pandas==1.5.3
pandocfilters==1.5.1
paramiko==3.5.1
parso==0.8.4
pathtools==0.1.2
pdfminer.six==20250327
pdfplumber==0.11.6
pexpect==4.9.0
Pillow==10.1.0
pip==25.1.1
platformdirs==4.3.8
playwright==1.52.0
ply==3.11
pooch==1.8.2
prometheus_client==0.22.0
promise==2.3
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
py-spy==0.4.0
py4j==0.10.9.9
pyacoustid==1.3.0
pyarrow==19.0.1
pyasn1==0.4.8
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.18.0
pycurl==7.43.0.6
pydantic==1.10.22
pydantic_core==2.33.2
PyDrive==1.3.1
pyee==13.0.0
PyGetWindow==0.0.9
Pygments==2.18.0
PyGObject==3.38.0
PyJWT==1.7.1
PyMsgBox==1.0.9
PyMuPDF==1.25.5
PyNaCl==1.5.0
pynput==1.7.8
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypdf==5.5.0
PyPDF2==3.0.1
pypdfium2==4.30.1
pyperclip==1.9.0
PyRect==0.2.0
pyrsistent==0.20.0
PyScreeze==1.0.1
PySocks==1.7.1
python-apt==2.2.1
python-barcode==0.15.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-etcd==0.4.5
python-jose==3.4.0
python-json-logger==3.3.0
python-pptx==1.0.2
python-snappy==0.7.3
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
pytz==2025.2
PyWavelets==1.6.0
PyYAML==6.0.2
pyzmq==27.0.1
qrcode==8.2
qtconsole==5.6.1
QtPy==2.4.3
RapidFuzz==3.13.0
reactivex==4.0.4
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.0.0
rpds-py==0.27.0
rsa==4.9.1
s3transfer==0.12.0
safetensors==0.5.3
schedula==1.5.62
schedule==1.2.2
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
Send2Trash==1.8.3
sentry-sdk==2.28.0
setproctitle==1.3.6
setuptools==65.7.0
shapely==2.0.7
shellescape==3.8.1
shortuuid==1.0.13
six==1.16.0
smart-open==6.4.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.6
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.8
sphinxcontrib-serializinghtml==1.1.10
sphinxcontrib-websupport==1.2.7
SQLAlchemy==2.0.27
stack-data==0.6.3
starlette==0.27.0
sympy==1.13.1
termcolor==3.1.0
terminado==0.18.1
textual==3.2.0
threadpoolctl==3.6.0
thrift==0.21.0
thriftpy2==0.5.2
tidebox==0.1.13
tifffile==2024.8.30
tiktoken==0.9.0
timm==1.0.7
tinycss2==1.4.0
tokenizers==0.15.2
tomli==2.2.1
torch==2.1.2
torchaudio==2.3.1+cu121
torchvision==0.16.2
tornado==6.5.2
tqdm==4.65.0
traitlets==5.14.3
transformers==4.35.2
triton==2.1.0
types-python-dateutil==2.9.0.20250822
typing_extensions==4.14.1
typing-inspection==0.4.0
tzdata==2025.2
uc-micro-py==1.0.3
ujson==5.10.0
unattended-upgrades==0.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.29.0
uvloop==0.21.0
virtualenv==20.31.2
wandb==0.19.11
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.43.0
widgetsnbextension==4.0.14
wrapt==1.17.2
wrapt_timeout_decorator==1.5.1
XlsxWriter==3.2.3
xmltodict==0.14.2
yarl==1.20.0
zipp==3.19.2
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
